+++ kohii-exoplayer/src/main/java/kohii/v1/exoplayer/PlayerViewBridgeCreator.kt
@@ -17,13 +17,15 @@
 package kohii.v1.exoplayer
 
 import android.content.Context
+import com.google.android.exoplayer2.Player
 import com.google.android.exoplayer2.ui.PlayerView
 import kohii.v1.core.Bridge
 import kohii.v1.core.BridgeCreator
+import kohii.v1.core.PlayerPool
 import kohii.v1.media.Media
 
 class PlayerViewBridgeCreator(
-  private val playerProvider: ExoPlayerProvider,
+  private val playerPool: PlayerPool<Player>,
   private val mediaSourceFactoryProvider: MediaSourceFactoryProvider
 ) : BridgeCreator<PlayerView> {
 
@@ -34,12 +36,12 @@ class PlayerViewBridgeCreator(
     return PlayerViewBridge(
         context,
         media,
-        playerProvider,
+        playerPool,
         mediaSourceFactoryProvider
     )
   }
 
   override fun cleanUp() {
-    playerProvider.cleanUp()
+    playerPool.clear()
   }
 }